backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.4.0
alembic==1.13.1

